Use local fee estimation for WC Update transactions
With the local estimation we are not restricted by GET request query
size limit and can handle quite big smart contract parameters.

+++ app/src/main/java/com/concordium/wallet/ui/cis2/SendTokenViewModel.kt
@@ -313,7 +313,7 @@ class SendTokenViewModel(application: Application) : AndroidViewModel(applicatio
             memoSize = if (sendTokenData.memo == null) null else sendTokenData.memo!!.length / 2,
             success = {
                 sendTokenData.energy = it.energy
-                sendTokenData.fee = it.cost.toBigInteger()
+                sendTokenData.fee = it.cost
                 sendTokenData.account?.let { account ->
                     sendTokenData.max =
                         account.getAtDisposalWithoutStakedOrScheduled(account.totalUnshieldedBalance) -
@@ -346,7 +346,7 @@ class SendTokenViewModel(application: Application) : AndroidViewModel(applicatio
             parameter = serializeTokenTransferParametersOutput.parameter,
             success = {
                 sendTokenData.energy = it.energy
-                sendTokenData.fee = it.cost.toBigInteger()
+                sendTokenData.fee = it.cost
                 waiting.postValue(false)
                 feeReady.postValue(sendTokenData.fee)
             },
